Move all action creators into their respective reducers.
