+++ src/components/Navigation/LanguageDropdown.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import {setLanguage, resetFilter} from '../../action_creators';
+import {setLanguage} from '../../reducers/language';
+import {resetFilter} from '../../reducers/filter';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import MenuItem from 'react-bootstrap/lib/MenuItem';
 import DropdownButton from 'react-bootstrap/lib/DropdownButton';
