+++ src/reducers/language.js
@@ -1,21 +1,32 @@
-/*
-  Example state:
-
-  language: "nb"
- */
-
 import {getAvailableLanguages} from '../util';
 import {storeItem} from '../localStorage';
 
-export default function(state='nb', action) {
+
+/////////////////////
+// ACTION CREATORS //
+/////////////////////
+
+export function setLanguage(language) {
+  return {
+    type: 'SET_LANGUAGE',
+    language
+  };
+}
+
+
+/////////////
+// REDUCER //
+/////////////
+
+const INITIAL_STATE = 'nb';
+
+const isLanguageValid = (language) => getAvailableLanguages().includes(language);
+
+export default function(state = INITIAL_STATE, action) {
   switch (action.type) {
     case 'SET_LANGUAGE':
       return isLanguageValid(action.language) ? storeItem('language', action.language) : state;
   }
 
   return state;
-}
-
-function isLanguageValid(language) {
-  return getAvailableLanguages().includes(language);
 }
