+++ src/reducers/welcomeBox.js
@@ -1,11 +1,25 @@
-/*
-  Example state:
-
-  welcomeBox: true
- */
 import {storeItem} from '../localStorage';
 
-export default function(state=true, action) {
+
+/////////////////////
+// ACTION CREATORS //
+/////////////////////
+
+export function setWelcomeBox(welcomeBox) {
+  return {
+    type: 'SET_WELCOMEBOX',
+    payload: welcomeBox
+  };
+}
+
+
+/////////////
+// REDUCER //
+/////////////
+
+const INITIAL_STATE = true;
+
+export default function(state = INITIAL_STATE, action) {
   switch(action.type) {
     case 'SET_WELCOMEBOX':
       return storeItem('welcomeBox', (action.payload));
